+++ src/test/java/com/github/yermaalexx/departmentsandemployees/unittests/departments/GetDepartmentUnitTests.java
@@ -1,4 +1,4 @@
-package com.github.yermaalexx.departmentsandemployees.unit_tests.departments;
+package com.github.yermaalexx.departmentsandemployees.unittests.departments;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -8,8 +8,10 @@ import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.modelmapper.ModelMapper;
 
+import com.github.yermaalexx.departmentsandemployees.entities.DepartmentEntity;
 import com.github.yermaalexx.departmentsandemployees.entities.EmployeeEntity;
 import com.github.yermaalexx.departmentsandemployees.exceptions.NoDepartmentWithThisIDException;
+import com.github.yermaalexx.departmentsandemployees.models.DepartmentDTO;
 import com.github.yermaalexx.departmentsandemployees.models.EmployeeDTO;
 import com.github.yermaalexx.departmentsandemployees.repositories.DepartmentRepository;
 import com.github.yermaalexx.departmentsandemployees.repositories.EmployeeRepository;
@@ -17,14 +19,15 @@ import com.github.yermaalexx.departmentsandemployees.services.DepartmentService;
 
 import java.time.LocalDate;
 import java.util.List;
+import java.util.Optional;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.BDDMockito.given;
 
 @ExtendWith(MockitoExtension.class)
-@DisplayName("Test support method employeeListOfDepartment() from DepartmentService.class")
-public class EmployeeListOfDepartmentUnitTests {
+@DisplayName("Test method getDepartment(Integer id) from DepartmentService.class")
+public class GetDepartmentUnitTests {
 
     @Mock
     private DepartmentRepository departmentRepository;
@@ -37,25 +40,30 @@ public class EmployeeListOfDepartmentUnitTests {
 
     @Test
     @DisplayName("Normal flow")
-    public void employeeListOfDepartmentNormal() {
+    public void getDepartmentNormal() {
         Integer id = 3;
         given(departmentRepository.existsById(id)).willReturn(true);
+        DepartmentEntity fromRepository = new DepartmentEntity(3, "Development department", "Some description", "Location: NYC");
+        DepartmentDTO fromMapper = new DepartmentDTO(3, "Development department", "Some description", "Location: NYC");
+        given(departmentRepository.findById(id)).willReturn(Optional.of(fromRepository));
+        given(modelMapper.map(fromRepository, DepartmentDTO.class)).willReturn(fromMapper);
         EmployeeDTO employeeDTO = new EmployeeDTO(1, "Musk Ilon", LocalDate.of(1999, 03, 17), LocalDate.of(2020, 04, 15), 3, "Manager");
         EmployeeEntity employeeEntity = new EmployeeEntity(1, "Musk Ilon", LocalDate.of(1999, 03, 17), LocalDate.of(2020, 04, 15), 3, "Manager");
         List<EmployeeEntity> employeeEntityList = List.of(employeeEntity);
         List<EmployeeDTO> employeeDTOList = List.of(employeeDTO);
         given(employeeRepository.findAll()).willReturn(employeeEntityList);
         given(modelMapper.map(employeeEntity, EmployeeDTO.class)).willReturn(employeeDTO);
-        List<EmployeeDTO> result = departmentService.employeeListOfDepartment(id);
-        assertEquals(employeeDTOList, result);
+        DepartmentDTO endDTO = new DepartmentDTO(3, "Development department", "Some description", "Location: NYC", employeeDTOList);
+        DepartmentDTO result = departmentService.getDepartment(id);
+        assertEquals(endDTO, result);
     }
 
     @Test
     @DisplayName("No department with this ID")
-    public void employeeListOfDepartmentIdIsIncorrect() {
+    public void getDepartmentIdIsIncorrect() {
         Integer id = 3;
         given(departmentRepository.existsById(id)).willReturn(false);
-        assertThrows(NoDepartmentWithThisIDException.class, () -> departmentService.employeeListOfDepartment(id));
+        assertThrows(NoDepartmentWithThisIDException.class, () -> departmentService.getDepartment(id));
     }
 
 }
