+++ src/test/java/com/github/yermaalexx/departmentsandemployees/unittests/departments/GetListOfDepartmentsUnitTests.java
@@ -1,4 +1,4 @@
-package com.github.yermaalexx.departmentsandemployees.unit_tests.departments;
+package com.github.yermaalexx.departmentsandemployees.unittests.departments;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -10,7 +10,6 @@ import org.modelmapper.ModelMapper;
 
 import com.github.yermaalexx.departmentsandemployees.entities.DepartmentEntity;
 import com.github.yermaalexx.departmentsandemployees.entities.EmployeeEntity;
-import com.github.yermaalexx.departmentsandemployees.exceptions.NoDepartmentWithThisIDException;
 import com.github.yermaalexx.departmentsandemployees.models.DepartmentDTO;
 import com.github.yermaalexx.departmentsandemployees.models.EmployeeDTO;
 import com.github.yermaalexx.departmentsandemployees.repositories.DepartmentRepository;
@@ -19,15 +18,13 @@ import com.github.yermaalexx.departmentsandemployees.services.DepartmentService;
 
 import java.time.LocalDate;
 import java.util.List;
-import java.util.Optional;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.BDDMockito.given;
 
 @ExtendWith(MockitoExtension.class)
-@DisplayName("Test method getDepartment(Integer id) from DepartmentService.class")
-public class GetDepartmentUnitTests {
+@DisplayName("Test method getListOfDepartments() from DepartmentService.class")
+public class GetListOfDepartmentsUnitTests {
 
     @Mock
     private DepartmentRepository departmentRepository;
@@ -40,13 +37,13 @@ public class GetDepartmentUnitTests {
 
     @Test
     @DisplayName("Normal flow")
-    public void getDepartmentNormal() {
-        Integer id = 3;
-        given(departmentRepository.existsById(id)).willReturn(true);
-        DepartmentEntity fromRepository = new DepartmentEntity(3, "Development department", "Some description", "Location: NYC");
-        DepartmentDTO fromMapper = new DepartmentDTO(3, "Development department", "Some description", "Location: NYC");
-        given(departmentRepository.findById(id)).willReturn(Optional.of(fromRepository));
-        given(modelMapper.map(fromRepository, DepartmentDTO.class)).willReturn(fromMapper);
+    public void getListOfDepartmentsNormal() {
+        DepartmentEntity departmentEntity = new DepartmentEntity(3, "Development department", "Some description", "Location: NYC");
+        DepartmentDTO departmentDTO = new DepartmentDTO(3, "Development department", "Some description", "Location: NYC");
+        List<DepartmentEntity> departmentEntityList = List.of(departmentEntity);
+        given(departmentRepository.findAll()).willReturn(departmentEntityList);
+        given(modelMapper.map(departmentEntity, DepartmentDTO.class)).willReturn(departmentDTO);
+        given(departmentRepository.existsById(3)).willReturn(true);
         EmployeeDTO employeeDTO = new EmployeeDTO(1, "Musk Ilon", LocalDate.of(1999, 03, 17), LocalDate.of(2020, 04, 15), 3, "Manager");
         EmployeeEntity employeeEntity = new EmployeeEntity(1, "Musk Ilon", LocalDate.of(1999, 03, 17), LocalDate.of(2020, 04, 15), 3, "Manager");
         List<EmployeeEntity> employeeEntityList = List.of(employeeEntity);
@@ -54,16 +51,12 @@ public class GetDepartmentUnitTests {
         given(employeeRepository.findAll()).willReturn(employeeEntityList);
         given(modelMapper.map(employeeEntity, EmployeeDTO.class)).willReturn(employeeDTO);
         DepartmentDTO endDTO = new DepartmentDTO(3, "Development department", "Some description", "Location: NYC", employeeDTOList);
-        DepartmentDTO result = departmentService.getDepartment(id);
-        assertEquals(endDTO, result);
+        List<DepartmentDTO> departmentDTOList = List.of(endDTO);
+        List<DepartmentDTO> result = departmentService.getListOfDepartments();
+        assertEquals(departmentDTOList, result);
     }
 
-    @Test
-    @DisplayName("No department with this ID")
-    public void getDepartmentIdIsIncorrect() {
-        Integer id = 3;
-        given(departmentRepository.existsById(id)).willReturn(false);
-        assertThrows(NoDepartmentWithThisIDException.class, () -> departmentService.getDepartment(id));
-    }
+
+
 
 }
