+++ src/test/java/com/github/yermaalexx/departmentsandemployees/unittests/departments/GetListOfEmployeesForDepartmentUnitTests.java
@@ -1,4 +1,4 @@
-package com.github.yermaalexx.departmentsandemployees.unit_tests.departments;
+package com.github.yermaalexx.departmentsandemployees.unittests.departments;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -8,9 +8,8 @@ import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.modelmapper.ModelMapper;
 
-import com.github.yermaalexx.departmentsandemployees.entities.DepartmentEntity;
 import com.github.yermaalexx.departmentsandemployees.entities.EmployeeEntity;
-import com.github.yermaalexx.departmentsandemployees.models.DepartmentDTO;
+import com.github.yermaalexx.departmentsandemployees.exceptions.NoDepartmentWithThisIDException;
 import com.github.yermaalexx.departmentsandemployees.models.EmployeeDTO;
 import com.github.yermaalexx.departmentsandemployees.repositories.DepartmentRepository;
 import com.github.yermaalexx.departmentsandemployees.repositories.EmployeeRepository;
@@ -20,11 +19,12 @@ import java.time.LocalDate;
 import java.util.List;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.BDDMockito.given;
 
 @ExtendWith(MockitoExtension.class)
-@DisplayName("Test method getListOfDepartments() from DepartmentService.class")
-public class GetListOfDepartmentsUnitTests {
+@DisplayName("Test method getListOfEmployeesForDepartment(Integer departmentID) from DepartmentService.class")
+public class GetListOfEmployeesForDepartmentUnitTests {
 
     @Mock
     private DepartmentRepository departmentRepository;
@@ -37,26 +37,25 @@ public class GetListOfDepartmentsUnitTests {
 
     @Test
     @DisplayName("Normal flow")
-    public void getListOfDepartmentsNormal() {
-        DepartmentEntity departmentEntity = new DepartmentEntity(3, "Development department", "Some description", "Location: NYC");
-        DepartmentDTO departmentDTO = new DepartmentDTO(3, "Development department", "Some description", "Location: NYC");
-        List<DepartmentEntity> departmentEntityList = List.of(departmentEntity);
-        given(departmentRepository.findAll()).willReturn(departmentEntityList);
-        given(modelMapper.map(departmentEntity, DepartmentDTO.class)).willReturn(departmentDTO);
-        given(departmentRepository.existsById(3)).willReturn(true);
+    public void getListOfEmployeesForDepartmentNormal() {
+        Integer id = 3;
+        given(departmentRepository.existsById(id)).willReturn(true);
         EmployeeDTO employeeDTO = new EmployeeDTO(1, "Musk Ilon", LocalDate.of(1999, 03, 17), LocalDate.of(2020, 04, 15), 3, "Manager");
         EmployeeEntity employeeEntity = new EmployeeEntity(1, "Musk Ilon", LocalDate.of(1999, 03, 17), LocalDate.of(2020, 04, 15), 3, "Manager");
         List<EmployeeEntity> employeeEntityList = List.of(employeeEntity);
         List<EmployeeDTO> employeeDTOList = List.of(employeeDTO);
         given(employeeRepository.findAll()).willReturn(employeeEntityList);
         given(modelMapper.map(employeeEntity, EmployeeDTO.class)).willReturn(employeeDTO);
-        DepartmentDTO endDTO = new DepartmentDTO(3, "Development department", "Some description", "Location: NYC", employeeDTOList);
-        List<DepartmentDTO> departmentDTOList = List.of(endDTO);
-        List<DepartmentDTO> result = departmentService.getListOfDepartments();
-        assertEquals(departmentDTOList, result);
+        List<EmployeeDTO> result = departmentService.getListOfEmployeesForDepartment(id);
+        assertEquals(employeeDTOList, result);
     }
 
-
-
+    @Test
+    @DisplayName("No department with this ID")
+    public void getListOfEmployeesForDepartmentIdIsIncorrect() {
+        Integer id = 3;
+        given(departmentRepository.existsById(id)).willReturn(false);
+        assertThrows(NoDepartmentWithThisIDException.class, () -> departmentService.getListOfEmployeesForDepartment(id));
+    }
 
 }
