+++ src/test/java/com/github/yermaalexx/departmentsandemployees/unittests/employees/DeleteEmployeeUnitTests.java
@@ -1,4 +1,4 @@
-package com.github.yermaalexx.departmentsandemployees.unit_tests.employees;
+package com.github.yermaalexx.departmentsandemployees.unittests.employees;
 
 import com.github.yermaalexx.departmentsandemployees.services.EmployeeService;
 import org.junit.jupiter.api.DisplayName;
@@ -7,48 +7,38 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
-import org.modelmapper.ModelMapper;
 
-import com.github.yermaalexx.departmentsandemployees.entities.EmployeeEntity;
 import com.github.yermaalexx.departmentsandemployees.exceptions.NoEmployeeWithThisIDException;
-import com.github.yermaalexx.departmentsandemployees.models.EmployeeDTO;
 import com.github.yermaalexx.departmentsandemployees.repositories.EmployeeRepository;
-
-import java.time.LocalDate;
-import java.util.Optional;
 
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.BDDMockito.given;
 import static org.mockito.Mockito.verify;
 
 @ExtendWith(MockitoExtension.class)
-@DisplayName("Test method getEmployee(Integer id) from EmployeeService.class")
-public class GetEmployeeUnitTests {
+@DisplayName("Test method deleteEmployee(Integer id) from EmployeeService.class")
+public class DeleteEmployeeUnitTests {
 
     @Mock
     private EmployeeRepository employeeRepository;
-    @Mock
-    private ModelMapper modelMapper;
     @InjectMocks
     private EmployeeService employeeService;
 
     @Test
     @DisplayName("Normal flow")
-    public void getEmployeeNormal() {
+    public void deleteEmployeeNormal() {
         int id = 3;
         given(employeeRepository.existsById(id)).willReturn(true);
-        EmployeeEntity fromRepository = new EmployeeEntity(3, "Musk Ilon", LocalDate.of(1999, 03, 17), LocalDate.of(2020, 04, 15), 2, "Manager");
-        given(employeeRepository.findById(id)).willReturn(Optional.of(fromRepository));
-        employeeService.getEmployee(id);
-        verify(modelMapper).map(fromRepository, EmployeeDTO.class);
+        employeeService.deleteEmployee(id);
+        verify(employeeRepository).deleteById(id);
     }
 
     @Test
     @DisplayName("No employee with this ID")
-    public void getEmployeeIdIsIncorrect() {
+    public void deleteEmployeeIdIsIncorrect() {
         int id = 3;
         given(employeeRepository.existsById(id)).willReturn(false);
-        assertThrows(NoEmployeeWithThisIDException.class, () -> employeeService.getEmployee(id));
+        assertThrows(NoEmployeeWithThisIDException.class, () -> employeeService.deleteEmployee(id));
     }
 
 }
